+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Added:
 
 * Compatibility with KissMyRank v1.6f
 * Added a "Public URL" option to sTracker's HTTP Configuration to be used when linking to sTracker from Live Timings. If you were having problems with parts of sTracker not working correctly, try setting the Public URL option.
+* STracker and KissMyRank now start and stop with Server Manager, not the Assetto Corsa Server, so you can access them all the time.
 
 Fixes: